pyproject.toml
[tool.poetry]
name = "crossstitchgenerator"
version = "0.1.0"
description = "Generates a cross stitch pattern from any image!"
authors = ["Karla Barraza <barraza.karla1234@gmail.com>"]
readme = "README.md"
package-mode = false 

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^2.0.0"
pillow = "^10.3.0"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
